Improve thread safety around Window a little bit
Window contains a linked list of windows through xrd_window's
child/parent link. Modifying this linked list could be dangerous if one
doesn't have exclusive access to WindowState.

Normally we do. But for example we remove a window from WindowState and
drop it, we could have safety problems if we aren't careful.

We make Window::drop unsafe and require it to be explicitly called to
avoid this problem.