put X cursor mirror on top of everything else
add XrdWindow::set_sort_order, and use it to set the sort order of the
cursor mirror.

OpenVR overlays doesn't use depth testing, they are just rendered in
order, based on their sort order. so the cursor just being in front of
the window is not enough, it must have a higher sort order.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>